Rename `$cam->{weight}`, `$set->weight` to `mass`
This is because the word "weight" has both a specific meaning for cams, and
(potentially) a more general meaning ("weighted average" and the like).

Similarly, when the code acquires the ability to optimise for minimal
financial expenditure, we'll use the word "price" rather than "cost" (which
would have a particular association with the terms "cost function" and "cost
method" that the code relies on).